requirements.txt
fastapi
uvicorn
python-multipart
supabase
python-dotenv
torch
transformers
pillow
scipy
tf-keras
